pyproject.toml
[tool.poetry]
name = "The simple chatbot"
version = "0.0.1"
authors = ["khaidq"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
gradio = "^5.13.2"
python-decouple = "^3.8"
ollama = "^0.4.7"
